test: controller manager to share with other tests
