+++ packages/integration-tests/projects/suite-web/plugins/index.js
@@ -6,6 +6,7 @@ import fs from 'fs';
 import path from 'path';
 import { addMatchImageSnapshotPlugin } from 'cypress-image-snapshot/plugin';
 import { Controller } from '@trezor/trezor-user-env-link';
+import controllerManager from '../../../controllerManager';
 import googleMock from './google';
 import dropboxMock from './dropbox';
 import bridgeMock from './bridge';
@@ -33,7 +34,9 @@ let port = 0;
 let client = null;
 
 const controller = new Controller();
-let firmwares = {};
+const firmwares = {};
+
+const manager = controllerManager(controller, firmwares);
 
 module.exports = on => {
     // make ts possible start
@@ -127,28 +130,6 @@ module.exports = on => {
                     throw new Error('not a valid case');
             }
         },
-        sendToAddressAndMineBlock: async options => {
-            await controller.send({
-                type: 'regtest-send-to-address',
-                ...options,
-            });
-            return null;
-        },
-        mineBlocks: async options => {
-            await controller.send({
-                type: 'regtest-mine-blocks',
-                ...options,
-            });
-            return null;
-        },
-        startBridge: async version => {
-            await controller.send({ type: 'bridge-start', version });
-            return null;
-        },
-        stopBridge: async () => {
-            const response = await controller.send({ type: 'bridge-stop' });
-            return null;
-        },
         startMockedBridge: async har => {
             await bridgeMock.start(har);
             return null;
@@ -157,102 +138,7 @@ module.exports = on => {
             await bridgeMock.stop();
             return null;
         },
-        setupEmu: async options => {
-            const defaults = {
-                // some random empty seed. most of the test don't need any account history so it is better not to slow them down with all all seed
-                mnemonic:
-                    'alcohol woman abuse must during monitor noble actual mixed trade anger aisle',
-                pin: '',
-                passphrase_protection: false,
-                label: 'My Trevor',
-                needs_backup: false,
-            };
 
-            // before setup, stop bridge and start it again after it. it has no performance hit
-            // and avoids 'wrong previous session' errors from bridge. actual setup is done
-            // through udp transport if bridge transport is not available
-            await controller.send({ type: 'bridge-stop' });
-            await controller.send({
-                type: 'emulator-setup',
-                ...defaults,
-                ...options,
-            });
-            await controller.send({ type: 'bridge-start' });
-            return null;
-        },
-        /**
-         * @version
-         * version of firmware in emulator, only few are supported
-         * @wipe
-         * shall be emulator wiped before start? defaults to true
-         */
-        startEmu: async arg => {
-            const params = {
-                type: 'emulator-start',
-                version: process.env.FIRMWARE || '2-latest',
-                ...arg,
-            };
-            await controller.send(params);
-            return params;
-        },
-        stopEmu: async () => {
-            await controller.send({ type: 'emulator-stop' });
-            return null;
-        },
-        wipeEmu: async () => {
-            await controller.send({ type: 'emulator-wipe' });
-            return null;
-        },
-        pressYes: async () => {
-            await controller.send({ type: 'emulator-press-yes' });
-            return null;
-        },
-        pressNo: async () => {
-            await controller.send({ type: 'emulator-press-no' });
-            return null;
-        },
-        swipeEmu: async direction => {
-            await controller.send({ type: 'emulator-swipe', direction });
-            return null;
-        },
-        inputEmu: async value => {
-            await controller.send({ type: 'emulator-input', value });
-            return null;
-        },
-        clickEmu: async options => {
-            await controller.send({ type: 'emulator-click', ...options });
-            return null;
-        },
-        resetDevice: async options => {
-            await controller.send({ type: 'emulator-reset-device', ...options });
-            return null;
-        },
-        readAndConfirmMnemonicEmu: async () => {
-            await controller.send({ type: 'emulator-read-and-confirm-mnemonic' });
-            return null;
-        },
-        readAndConfirmShamirMnemonicEmu: async options => {
-            await controller.send({
-                type: 'emulator-read-and-confirm-shamir-mnemonic',
-                ...options,
-            });
-            return null;
-        },
-        applySettings: async options => {
-            await controller.send({
-                type: 'emulator-apply-settings',
-                ...options,
-            });
-            return null;
-        },
-        selectNumOfWordsEmu: async num => {
-            await controller.send({ type: 'emulator-select-num-of-words', num });
-            return null;
-        },
-        logTestDetails: async text => {
-            await controller.send({ type: 'log', text });
-            return null;
-        },
         resetCRI: async () => {
             if (client) {
                 await client.close();
@@ -295,19 +181,6 @@ module.exports = on => {
             }
             return null;
         },
-        trezorUserEnvConnect: async () => {
-            await controller.connect();
-
-            controller.on('firmwares', data => {
-                firmwares = data;
-            });
-
-            return null;
-        },
-        trezorUserEnvDisconnect: async () => {
-            await controller.disconnect();
-            return null;
-        },
         csvToJson(data) {
             const lines = data.split('\n');
             const result = [];
@@ -323,5 +196,7 @@ module.exports = on => {
             }
             return result;
         },
+        // Here `manager` adds all the required helper for websocket controller.
+        ...manager,
     });
 };
